Data Asociations between modesl: - User/Cooment - User/ Camp
