+++ models/campground.js
@@ -4,6 +4,13 @@ const campgroundSchema = new mongoose.Schema({
     name: String,
     image: String,
     description: String,
+    author: {
+        id:{
+            type: mongoose.Schema.Types.ObjectId,
+            ref: "User"
+        },
+        username: String
+    },
     comments:[
         {
             type: mongoose.Schema.Types.ObjectId,
