+++ views/campgrounds/show.ejs
@@ -19,6 +19,7 @@
                     <h4 class="pull-right">$10.00 per/night</h4>
                     <h4><a href=""> <%= campground.name %> </a></h4>
                     <p><%= campground.description %> </p>
+                    <p><em>Submitted by: <%= campground.author.username %></em>  </p>
                 </div>
             </div>
             <div class="well">
@@ -29,7 +30,7 @@
                 <% campground.comments.forEach((comment)=>{ %>
                     <div class="row"> 
                         <div class="col-md-12">
-                            <strong><%= comment.author %></strong>
+                            <strong><%= comment.author.username %></strong>
                             
                             <span class="pull-right"> 10 days ago</span>
                             <p>
