+++ views/comments/new.ejs
@@ -8,9 +8,6 @@
                     <input class="form-control" type="text" name="comment[text]" placeholder="Text">
                 </div>
                 <div class="form-group">
-                    <input class="form-control" type="text" name="comment[author]" placeholder="author">
-                </div>
-                <div class="form-group">
                     <button class="btn btn-lg btn-primary btn-block">Submit !</button>
                 </div>                
             </form>
